RavenDB-25939 Add download attachment button in message
